# utils: re-implement and improve column-wise output