I'd like to work on `puts_columns` as I believe decoupling this from an external tool will open up new possibilities without significant performance loss or increase in complexity.

***Please continue reading in the [3rd comment from 2015-10-03](#issuecomment-145188663). The remainder of this initial commit does no longer apply, but is retained in case someone already read it.***

----

The implementation currently doesn't try to faithfully recreate the previous output. In fact, these are some (major) differences:

- Items ordered line by line from left to right (was column by column from top to bottom).
- No pagination (was 55 lines per “page” irrespective of terminal height).
- Full justification even for single-line output (was flush left).
- Reduced right margin (was always at least two spaces).

----

Pagination and item order are probably worth some discussion. Here are some arguments:

### Current: Top to Bottom, With Pagination

- Columns are easier to follow visually.
- Scrolling up and down is inconvenient (for columns taller than terminal height).
- Pagination is awkward as there is no indication when a new “page” starts (only affects long listings with more than 55 lines).
- Disagrees with left-to-right reading direction in English.
- Slightly harder to implement.

Example layout on a hypothetical 9-column terminal (`|` = boundaries, `.` = blanks):

```
|a..d..g..|
|b..e.....|
|c..f.....|
```

### Proposed: Left to Right, Without Pagination

- Lines are slightly harder to follow visually.
- No pagination, no back-and-forth scrolling while scanning the listing from front to back.
- Agrees with left-to-right reading direction in English.
- Easier to implement.

Example layout on a hypothetical 9-column terminal (`|` = boundaries, `.` = blanks):

```
|a...b...c|
|d...e...f|
|g........|
```

----

Looking forward to your feedback!

*Some context: My motivation for re-implementing this in pure Ruby is that I'd like to be able to highlight individual items of the listing using `Tty`'s escape codes. This isn't possible with the current approach using an external command, because it does not have support for escape codes, effectively destroying the column layout as soon as those are used. Since highlighting is an entirely new feature, I'd like to postpone the implementation until this PR gets merged. Ultimately, #44335 will benefit from this highlighting.*