@UniqMartin 2015-09-27
Thinking a bit more about it, I could imagine that the pagination done by `pr` wasn't actually intended and went unnoticed. If top to bottom turns out to be the preferred layout, I'd suggest to implement it without pagination, resembling the output of `ls`. (Though this will still require scrolling back and forth to scan the entire output of a long list in its original order.)