@UniqMartin 2015-10-03
There are two commits in this PR:

## 1: Reimplementation of Multi-Column Output

(Please review this first, as this is the primary topic of this PR. Ideally, I'd like to have a consensus here before tweaking the second part.)

The current implementation doesn't fully recreate the original output. This is intentional. Differences are:

- No pagination (was 55 lines per “page” irrespective of terminal height; default of `pr` command).
- Full justification even for single-line output (was flush left).
- Reduced right margin (was always at least two spaces).
- If environment variable`HOMEBREW_PUTS_COLUMNS_ROW_MAJOR` is set, put items in row-major order (new alternative). If unset, use column-major order (as before).
  - *Note:* I implemented both variants for testing purposes. I will remove one of them once there's a consensus on the preferred item arrangement.

## 2: Highlighting of Individual Items

(This part could also be moved into a separate PR once the first part gets accepted.)

Provided to make the code structure of the first part clearer; plus it conveys the bigger picture. The ultimate goal is to allow for highlighting of individual items in the multi-column listing. @xu-cheng suggested to implement this highlighting feature in a [diff comment](https://github.com/Homebrew/homebrew/pull/44335#discussion_r40449204) in #44335. I imagine @dunn might also be interested, according to comments [here](https://github.com/Homebrew/homebrew/pull/42662#issuecomment-129078548) and [here](https://github.com/Homebrew/homebrew/pull/44335#issuecomment-143476101). I wouldn't be surprised if there are even more uses for this feature.

I changed the interpretation of the second parameter of `puts_columns` for this. With this change it shouldn't be used to convey important information, as it will be only rendered on a TTY. However, this should be no problem, as there's currently no code in Homebrew core using this second parameter.