@UniqMartin 2015-10-03
Here's a screenshot that shows how highlighting could be used to enrich the output of `brew update`:

<img width="850" alt="Highlighting of installed formulae in formula update report." src="https://cloud.githubusercontent.com/assets/11892150/10260566/40e286e2-6978-11e5-9825-27a6abababcb.png">