@xu-cheng 2015-10-06
I wonder do you have any benchmark to show new implementation's performance vs old one.

cc @Homebrew/owners for advice