@UniqMartin 2015-10-06
> I wonder do you have any benchmark to show new implementation's performance vs old one.

I don't. It shouldn't be too difficult to benchmark this, though. I'll see what I can do and will report back.

However, my gut feeling is that it doesn't really matter, because almost always, the cost to generate the data to be output via `puts_columns` will be significantly higher (`brew update` probably being the most obvious example), making the performance of `puts_columns` itself negligible.