@UniqMartin 2015-10-06
> It shouldn't be too difficult to benchmark this, though. I'll see what I can do and will report back.

Here are benchmark results using `Benchmark.measure`. The external command that generated the results, the JSON test data, and outputs (on a terminal with 120 columns) can be found [in a gist](https://gist.github.com/UniqMartin/859f635c2d9513442958).

The two test files `data-{3,4}.json` are more or less what would be reported by `brew update` if the last update was 1000/10000 commits ago. (Note for comparison with the times below, that it took 22/136 seconds to generate these files.)

I ran every command three times and picked the best result. The output is user/system/total CPU time and (in parentheses) elapsed real time.

Old (non-native) code:

```
$ brew benchmark-multicol dump-X.json # on branch master
  0.010000   0.010000   0.060000 (  0.055237) - dump-3, non-native
  0.010000   0.010000   0.070000 (  0.073264) - dump-4, non-native
```

New (native) code, column-major order:

```
$ brew benchmark-multicol --highlight dump-X.json # on PR branch
  0.000000   0.000000   0.020000 (  0.020293) - dump-3, column-major, highlighted
  0.020000   0.010000   0.050000 (  0.052954) - dump-4, column-major, highlighted
```

New (native) code, row-major order:

```
$ brew benchmark-multicol --rows --highlight dump-X.json # on PR branch
  0.000000   0.000000   0.020000 (  0.019649) - dump-3, row-major, highlighted
  0.020000   0.010000   0.050000 (  0.052012) - dump-4, row-major, highlighted
```