@mikemcquaid 2015-10-13
This is sounding good to me so far. Can you suggest some commands I can run to view the difference in output? @xu-cheng any more thoughts? Thanks!