@xu-cheng 2015-10-13
This PR should be merged alongside with #44335.

LGTM, except we should decide what layout we should use.