@UniqMartin 2015-10-13
> Can you suggest some commands I can run to view the difference in output?

Try (prefix any of them with `HOMEBREW_PUTS_COLUMNS_ROW_MAJOR=1` for the row-major layout):

- `brew commands`,
- `brew list --full-name` (the only variant that uses `puts_columns`),
- `brew search`, `brew search pi` (or similar),
- `brew update` (probably easier to test with `brew benchmark-multicol` [from above](#issuecomment-145880361)), and
- `brew uses libpng` (or similar).

You can also use `brew irb` and then invoke `puts_columns` with pretty much any array of strings you can come up with.

> LGTM, except we should decide what layout we should use.

Indeed! For what it's worth, I'm now pretty much in favor of the (default) column-major layout.