@mikemcquaid 2015-10-14
Implementation and local testing look good. I also prefer the default column-major layout. Can you kill the `HOMEBREW_PUTS_COLUMNS_ROW_MAJOR` stuff and I'll merge? Thanks!