@UniqMartin 2015-10-14
Here we go. Just ripped out the `HOMEBREW_PUTS_COLUMNS_ROW_MAJOR` stuff, leaving column-major layout as the only option.